Module radix_engine_queries::typed_native_events

This module contains the code and models that are required to convert native events into a typed model based on the EventTypeIdentifier and the raw SBOR bytes of the event. This is used in the toolkit and consumed by the gateway for some of its internal operations.

Macros

define_structure 🔒

This enum uses some special syntax to define the structure of events. This makes the code for model definitions very compact, allows for very easy addition of more packages, blueprints or events in the future, keeps various models all in sync, and implements various functions and methods on appropriate types.

Enums

TypedAccessControllerBlueprintEvent

TypedAccessControllerBlueprintEventKey

TypedAccessControllerPackageEvent

TypedAccessControllerPackageEventKey

TypedAccountBlueprintEvent

TypedAccountBlueprintEventKey

TypedAccountPackageEvent

TypedAccountPackageEventKey

TypedComponentRoyaltyBlueprintEvent

TypedComponentRoyaltyBlueprintEventKey

TypedConsensusManagerBlueprintEvent

TypedConsensusManagerBlueprintEventKey

TypedConsensusManagerPackageEvent

TypedConsensusManagerPackageEventKey

TypedFungibleResourceManagerBlueprintEvent

TypedFungibleResourceManagerBlueprintEventKey

TypedFungibleVaultBlueprintEvent

TypedFungibleVaultBlueprintEventKey

TypedIdentityBlueprintEvent

TypedIdentityBlueprintEventKey

TypedIdentityPackageEvent

TypedIdentityPackageEventKey

TypedMetadataBlueprintEvent

TypedMetadataBlueprintEventKey

TypedMetadataPackageEvent

TypedMetadataPackageEventKey

TypedMultiResourcePoolBlueprintEvent

TypedMultiResourcePoolBlueprintEventKey

TypedNativeEvent

TypedNativeEventError

TypedNativeEventKey

TypedNonFungibleResourceManagerBlueprintEvent

TypedNonFungibleResourceManagerBlueprintEventKey

TypedNonFungibleVaultBlueprintEvent

TypedNonFungibleVaultBlueprintEventKey

TypedOneResourcePoolBlueprintEvent

TypedOneResourcePoolBlueprintEventKey

TypedPackageBlueprintEvent

TypedPackageBlueprintEventKey

TypedPackagePackageEvent

TypedPackagePackageEventKey

TypedPoolPackageEvent

TypedPoolPackageEventKey

TypedResourcePackageEvent

TypedResourcePackageEventKey

TypedRoleAssignmentBlueprintEvent

TypedRoleAssignmentBlueprintEventKey

TypedRoleAssignmentPackageEvent

TypedRoleAssignmentPackageEventKey

TypedRoyaltyPackageEvent

TypedRoyaltyPackageEventKey

TypedTransactionProcessorBlueprintEvent

TypedTransactionProcessorBlueprintEventKey

TypedTransactionProcessorPackageEvent

TypedTransactionProcessorPackageEventKey

TypedTransactionTrackerBlueprintEvent

TypedTransactionTrackerBlueprintEventKey

TypedTransactionTrackerPackageEvent

TypedTransactionTrackerPackageEventKey

TypedTwoResourcePoolBlueprintEvent

TypedTwoResourcePoolBlueprintEventKey

TypedValidatorBlueprintEvent

TypedValidatorBlueprintEventKey

Functions

resolve_typed_event_key_from_event_type_identifier 🔒

to_typed_event_with_event_key 🔒

to_typed_native_event

Given an EventTypeIdentifier and the raw event data, this function attempts to convert the event data into a structured model provided that the event is registered to a native blueprint.

Type Aliases

AccountAddAuthorizedDepositorEvent 🔒

AccountDepositEvent 🔒

AccountRejectedDepositEvent 🔒

AccountRemoveAuthorizedDepositorEvent 🔒

AccountRemoveResourcePreferenceEvent 🔒

AccountSetDefaultDepositRuleEvent 🔒

AccountSetResourcePreferenceEvent 🔒

AccountWithdrawEvent 🔒

FungibleVaultDepositEvent 🔒

FungibleVaultLockFeeEvent 🔒

FungibleVaultPayFeeEvent 🔒

FungibleVaultRecallEvent 🔒

FungibleVaultWithdrawEvent 🔒

MultiResourcePoolContributionEvent 🔒

MultiResourcePoolDepositEvent 🔒

MultiResourcePoolRedemptionEvent 🔒

MultiResourcePoolWithdrawEvent 🔒

NonFungibleVaultDepositEvent 🔒

NonFungibleVaultRecallEvent 🔒

NonFungibleVaultWithdrawEvent 🔒

OneResourcePoolContributionEvent 🔒

OneResourcePoolDepositEvent 🔒

OneResourcePoolRedemptionEvent 🔒

OneResourcePoolWithdrawEvent 🔒

TwoResourcePoolContributionEvent 🔒

TwoResourcePoolDepositEvent 🔒

TwoResourcePoolRedemptionEvent 🔒

TwoResourcePoolWithdrawEvent 🔒