Macro radix_engine_queries::typed_native_events::define_structure

macro_rules! define_structure {
    (
        $(
            $package_ident: ident => {
                $(
                    $blueprint_ident: ident => [
                        $($event_ty: ty),* $(,)?
                    ]
                ),* $(,)?
            }
        ),* $(,)?
    ) => { ... };
}

This enum uses some special syntax to define the structure of events. This makes the code for model definitions very compact, allows for very easy addition of more packages, blueprints or events in the future, keeps various models all in sync, and implements various functions and methods on appropriate types.

The syntax allowed for by this macro looks like the following:

define_structure! {
    package_name1 => {
        blueprint_name1 => [
            Event1,
            Event2,
            Event3,
        ],
        blueprint_name2 => [
            Event1,
        ]
    },
    package_name2 => {
        blueprint_name1 => [
            Event1,
        ],
        blueprint_name2 => [
            Event1,
            Event2,
        ]
    }
}