Module radix_engine_queries::typed_substate_layout

Modules

fungible_vault

multi_resource_pool

non_fungible_vault

one_resource_pool

two_resource_pool

Structs

AccessControllerBlueprint

AccessControllerFeatureSet

AccessControllerNativePackage

AccessControllerStateFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

AccessControllerStateSchemaInit

AccessControllerSubstate

AccountBlueprint

AccountNativePackage

ActiveValidatorSet

AuthConfig

AuthZone

AuthZoneBlueprint

BadgeWithdrawEvent

BlueprintDefinition

BlueprintDefinitionInit

BlueprintDependencies

BlueprintInterface

BlueprintVersion

BlueprintVersionKey

BurnFungibleResourceEvent

BurnNonFungibleResourceEvent

CancelBadgeWithdrawAttemptEvent

CancelRecoveryProposalEvent

CanonicalBlueprintId

ClaimXrdEvent

CodeHash

Represents a particular instance of code under a package

ComponentClaimRoyaltiesInput

ComponentRoyaltyAccumulatorFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ComponentRoyaltyBlueprint

ComponentRoyaltyCreateInput

ComponentRoyaltyFeatureSet

ComponentRoyaltyLockInput

ComponentRoyaltyMethodAmountEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ComponentRoyaltyMethodAmountKeyPayload

This new type represents the payload of the key of a particular collection.

ComponentRoyaltySetInput

ComponentRoyaltyStateSchemaInit

ComponentRoyaltySubstate

ConsensusManagerBlueprint

ConsensusManagerConfigSubstate

ConsensusManagerConfigurationFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ConsensusManagerCurrentProposalStatisticFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ConsensusManagerCurrentValidatorSetFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ConsensusManagerFeatureSet

ConsensusManagerNativePackage

ConsensusManagerProposerMilliTimestampFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ConsensusManagerProposerMinuteTimestampFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ConsensusManagerRegisteredValidatorByStakeEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ConsensusManagerRegisteredValidatorByStakeKeyPayload

This type represents the payload of the key of a particular sorted index collection.

ConsensusManagerSecondsPrecisionNativeCode

ConsensusManagerStateFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ConsensusManagerStateSchemaInit

ConsensusManagerSubstate

ConsensusManagerValidatorRewardsFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

CurrentProposalStatisticSubstate

CurrentValidatorSetSubstate

DroppedFungibleBucket

DroppedNonFungibleBucket

EpochChangeEvent

FunctionSchema

FungibleBucketBlueprint

FungibleProofBlueprint

FungibleProofSubstate

FungibleResourceManagerBlueprint

FungibleResourceManagerDivisibilityFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

FungibleResourceManagerFeatureSet

FungibleResourceManagerStateSchemaInit

FungibleResourceManagerTotalSupplyFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

FungibleVaultBalanceFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

FungibleVaultBlueprint

FungibleVaultFeatureSet

FungibleVaultFreezeStatusFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

FungibleVaultLockedBalanceFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

FungibleVaultStateSchemaInit

IndexedStateSchema

InitiateBadgeWithdrawAttemptEvent

InitiateRecoveryEvent

LockOwnerRoleEvent

LockPrimaryRoleEvent

MetadataEntryEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

MetadataEntryKeyPayload

This new type represents the payload of the key of a particular collection.

MetadataFeatureSet

MetadataNativePackage

MetadataStateSchemaInit

MintFungibleResourceEvent

MintNonFungibleResourceEvent

NonFungibleBucketBlueprint

NonFungibleProofBlueprint

NonFungibleProofSubstate

NonFungibleResourceManagerBlueprint

NonFungibleResourceManagerDataEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

NonFungibleResourceManagerDataKeyPayload

This new type represents the payload of the key of a particular collection.

NonFungibleResourceManagerFeatureSet

NonFungibleResourceManagerIdTypeFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

NonFungibleResourceManagerMutableFieldsFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

NonFungibleResourceManagerMutableFieldsV1

NonFungibleResourceManagerStateSchemaInit

NonFungibleResourceManagerTotalSupplyFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

NonFungibleVaultBalanceFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

NonFungibleVaultBlueprint

NonFungibleVaultFeatureSet

NonFungibleVaultFreezeStatusFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

NonFungibleVaultLockedResourceFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

NonFungibleVaultNonFungibleEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

NonFungibleVaultNonFungibleKeyPayload

This new type represents the payload of the key of a particular collection.

NonFungibleVaultStateSchemaInit

OwnerRoleSubstate

PackageAuthNativeBlueprint

PackageBlueprintVersionAuthConfigEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

PackageBlueprintVersionAuthConfigKeyPayload

This new type represents the payload of the key of a particular collection.

PackageBlueprintVersionDefinitionEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

PackageBlueprintVersionDefinitionKeyPayload

This new type represents the payload of the key of a particular collection.

PackageBlueprintVersionDependenciesEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

PackageBlueprintVersionDependenciesKeyPayload

This new type represents the payload of the key of a particular collection.

PackageBlueprintVersionRoyaltyConfigEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

PackageBlueprintVersionRoyaltyConfigKeyPayload

This new type represents the payload of the key of a particular collection.

PackageClaimRoyaltiesInput

PackageCodeInstrumentedCodeEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

PackageCodeInstrumentedCodeKeyPayload

This new type represents the payload of the key of a particular collection.

PackageCodeInstrumentedCodeV1

PackageCodeOriginalCodeEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

PackageCodeOriginalCodeKeyPayload

This new type represents the payload of the key of a particular collection.

PackageCodeOriginalCodeV1

PackageCodeVmTypeEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

PackageCodeVmTypeKeyPayload

This new type represents the payload of the key of a particular collection.

PackageCodeVmTypeV1

PackageDefinition

PackageExport

PackageFeatureSet

PackageNativePackage

PackageOwnerBadgeData

PackagePublishNativeInput

PackagePublishNativeManifestInput

PackagePublishWasmAdvancedInput

PackagePublishWasmAdvancedManifestInput

PackagePublishWasmInput

PackagePublishWasmManifestInput

PackageRoyaltyAccumulatorFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

PackageRoyaltyAccumulatorV1

PackageRoyaltyNativeBlueprint

PackageSchemaEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

PackageSchemaKeyPayload

This new type represents the payload of the key of a particular collection.

PackageStateSchemaInit

PackageStructure

ProofMoveableSubstate

ProposalStatistic

ProposerMilliTimestampSubstate

ProposerMinuteTimestampSubstate

ProtocolUpdateReadinessSignalEvent

RegisterValidatorEvent

RemoveMetadataEvent

ResourceNativePackage

RoleAssignmentAccessRuleEntryPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

RoleAssignmentAccessRuleKeyPayload

This new type represents the payload of the key of a particular collection.

RoleAssignmentFeatureSet

RoleAssignmentNativePackage

RoleAssignmentOwnerFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

RoleAssignmentStateSchemaInit

RoundChangeEvent

RoyaltyNativePackage

RoyaltyUtil

RuleSetUpdateEvent

SetMetadataEvent

SetOwnerRoleEvent

SetRoleEvent

StakeEvent

StaticRoleDefinition

StopTimedRecoveryEvent

TransactionTrackerBlueprint

TransactionTrackerCreateInput

TransactionTrackerCreateManifestInput

TransactionTrackerNativePackage

TransactionTrackerSubstateV1

TypeInfoBlueprint

UnlockPrimaryRoleEvent

UnregisterValidatorEvent

UnstakeData

UnstakeEvent

UpdateAcceptingStakeDelegationStateEvent

Validator

ValidatorBlueprint

ValidatorByStakeKey

ValidatorEmissionAppliedEvent

ValidatorFeatureSet

ValidatorFeeChangeRequest

ValidatorOwnerBadgeData

ValidatorProtocolUpdateReadinessSignalFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ValidatorProtocolUpdateReadinessSignalSubstate

ValidatorRewardAppliedEvent

ValidatorRewardsSubstate

ValidatorStateFieldPayload

This new type represents the payload of a particular field or collection. It is unique to this particular field/collection.

ValidatorStateSchemaInit

ValidatorSubstate

VaultCreationEvent

VaultUtil

WorktopBlueprint

WorktopSubstate

Enums

AccessControllerCollection

AccessControllerError

AccessControllerFeature

AccessControllerField

AccessControllerTypedFieldSubstateValue

AccessControllerTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

AccessControllerTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

AccountError

AuthZoneError

BlueprintPartitionType

BlueprintPayloadDef

BlueprintPayloadIdentifier

BlueprintType

BootLoaderSubstateValue

BucketError

ComponentRoyaltyCollection

ComponentRoyaltyError

ComponentRoyaltyFeature

ComponentRoyaltyField

ComponentRoyaltyTypedFieldSubstateValue

ComponentRoyaltyTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

ComponentRoyaltyTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

ComposeProofError

ComposedProof

ConsensusManagerCollection

ConsensusManagerError

ConsensusManagerFeature

ConsensusManagerField

ConsensusManagerTypedFieldSubstateValue

ConsensusManagerTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

ConsensusManagerTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

FunctionAuth

FungibleResourceManagerCollection

FungibleResourceManagerError

Represents an error when accessing a bucket.

FungibleResourceManagerFeature

FungibleResourceManagerField

FungibleResourceManagerTypedFieldSubstateValue

FungibleResourceManagerTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

FungibleResourceManagerTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

FungibleVaultCollection

FungibleVaultFeature

FungibleVaultField

FungibleVaultTypedFieldSubstateValue

FungibleVaultTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

FungibleVaultTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

GenericScryptoComponentFieldValue

InputOrOutput

InvalidNonFungibleSchema

Represents an error when accessing a bucket.

KeyOrValue

LocalRef

MetadataCollection

MetadataError

MetadataFeature

MetadataField

MetadataTypedFieldSubstateValue

MetadataTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

MetadataTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

MetadataValidationError

MethodAuthTemplate

NonFungibleResourceManagerCollection

NonFungibleResourceManagerError

Represents an error when accessing a bucket.

NonFungibleResourceManagerFeature

NonFungibleResourceManagerField

NonFungibleResourceManagerGeneric

NonFungibleResourceManagerTypedFieldSubstateValue

NonFungibleResourceManagerTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

NonFungibleResourceManagerTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

NonFungibleVaultCollection

NonFungibleVaultError

NonFungibleVaultFeature

NonFungibleVaultField

NonFungibleVaultTypedFieldSubstateValue

NonFungibleVaultTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

NonFungibleVaultTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

PackageCollection

PackageError

PackageFeature

PackageField

PackageTypedFieldSubstateValue

PackageTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

PackageTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

PartitionDescription

PrimaryRoleBadgeWithdrawAttemptState

PrimaryRoleLockingState

PrimaryRoleRecoveryAttemptState

ProofError

RecoveryRoleBadgeWithdrawAttemptState

RecoveryRoleRecoveryAttemptState

RecoveryRoleRecoveryState

RoleAssignmentCollection

RoleAssignmentError

RoleAssignmentFeature

RoleAssignmentField

RoleAssignmentTypedFieldSubstateValue

RoleAssignmentTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

RoleAssignmentTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

RoleSpecification

SystemInstruction

TransactionStatus

TransactionStatusV1

TransactionTrackerSubstate

TypeInfoSubstate

TypedBootLoaderSubstateKey

TypedMainModuleSubstateKey

Doesn’t include non-object modules, nor transient nodes.

TypedMainModuleSubstateValue

Contains all the main module substate values, by each known partition layout

TypedMetadataModuleSubstateKey

TypedMetadataModuleSubstateValue

TypedRoleAssignmentModuleSubstateValue

TypedRoleAssignmentSubstateKey

TypedRoyaltyModuleSubstateKey

TypedRoyaltyModuleSubstateValue

TypedSchemaSubstateKey

TypedSubstateKey

TypedSubstateValue

TypedTransactionTrackerFieldValue

TypedTypeInfoModuleSubstateValue

TypedTypeInfoSubstateKey

ValidatorCollection

ValidatorError

ValidatorFeature

ValidatorField

ValidatorTypedFieldSubstateValue

ValidatorTypedSubstateKey

All the SubstateKeys for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

ValidatorTypedSubstateValue

All the Substate values for all logical partitions for the $blueprint_ident blueprint. Does not include mapped partitions, as these substates are mapped via their canonical partition.

VaultError

VersionedAccessControllerState

VersionedComponentRoyaltyAccumulator

VersionedComponentRoyaltyMethodAmount

VersionedConsensusManagerConfiguration

VersionedConsensusManagerCurrentProposalStatistic

VersionedConsensusManagerCurrentValidatorSet

VersionedConsensusManagerProposerMilliTimestamp

VersionedConsensusManagerProposerMinuteTimestamp

VersionedConsensusManagerRegisteredValidatorByStake

VersionedConsensusManagerState

VersionedConsensusManagerValidatorRewards

VersionedFungibleResourceManagerDivisibility

VersionedFungibleResourceManagerTotalSupply

VersionedFungibleVaultBalance

VersionedFungibleVaultFreezeStatus

VersionedFungibleVaultLockedBalance

VersionedMetadataEntry

VersionedNonFungibleResourceManagerIdType

VersionedNonFungibleResourceManagerMutableFields

VersionedNonFungibleResourceManagerTotalSupply

VersionedNonFungibleVaultBalance

VersionedNonFungibleVaultFreezeStatus

VersionedNonFungibleVaultLockedResource

VersionedNonFungibleVaultNonFungible

VersionedPackageBlueprintVersionAuthConfig

VersionedPackageBlueprintVersionDefinition

VersionedPackageBlueprintVersionDependencies

VersionedPackageBlueprintVersionRoyaltyConfig

VersionedPackageCodeInstrumentedCode

VersionedPackageCodeOriginalCode

VersionedPackageCodeVmType

VersionedPackageRoyaltyAccumulator

VersionedRoleAssignmentAccessRule

VersionedRoleAssignmentOwner

VersionedValidatorProtocolUpdateReadinessSignal

VersionedValidatorState

VmType

WorktopError

Constants

ACCESS_CONTROLLER_CODE_ID

ACCOUNT_CODE_ID

COMPONENT_ROYALTY_BLUEPRINT

COMPONENT_ROYALTY_CLAIMER_ROLE

COMPONENT_ROYALTY_CLAIMER_UPDATER_ROLE

COMPONENT_ROYALTY_CLAIM_ROYALTIES_IDENT

COMPONENT_ROYALTY_CREATE_IDENT

COMPONENT_ROYALTY_LOCKER_ROLE

COMPONENT_ROYALTY_LOCKER_UPDATER_ROLE

COMPONENT_ROYALTY_LOCK_ROYALTY_IDENT

COMPONENT_ROYALTY_SETTER_ROLE

COMPONENT_ROYALTY_SETTER_UPDATER_ROLE

COMPONENT_ROYALTY_SET_ROYALTY_IDENT

CONSENSUS_MANAGER_CODE_ID

CONSENSUS_MANAGER_REGISTERED_VALIDATORS_BY_STAKE_INDEX

CONSENSUS_MANAGER_SECONDS_PRECISION_CODE_ID

EPOCHS_PER_PARTITION

IDENTITY_CODE_ID

METADATA_CODE_ID

OWNER_STAKE_UNITS_PENDING_WITHDRAWALS_LIMIT

A performance-driven limit on the number of simultaneously pending “delayed withdrawal” operations on any validator’s owner’s stake units vault.

PACKAGE_AUTH_TEMPLATE_PARTITION_OFFSET

PACKAGE_BLUEPRINT

PACKAGE_BLUEPRINTS_PARTITION_OFFSET

PACKAGE_BLUEPRINT_DEPENDENCIES_PARTITION_OFFSET

PACKAGE_CLAIM_ROYALTIES_IDENT

PACKAGE_CODE_ID

PACKAGE_FIELDS_PARTITION_OFFSET

PACKAGE_INSTRUMENTED_CODE_PARTITION_OFFSET

PACKAGE_ORIGINAL_CODE_PARTITION_OFFSET

PACKAGE_PUBLISH_NATIVE_IDENT

PACKAGE_PUBLISH_WASM_ADVANCED_IDENT

PACKAGE_PUBLISH_WASM_IDENT

PACKAGE_ROYALTY_PARTITION_OFFSET

PACKAGE_VM_TYPE_PARTITION_OFFSET

PARTITION_RANGE_END

PARTITION_RANGE_START

POOL_V1_0_CODE_ID

POOL_V1_1_CODE_ID

RESOURCE_CODE_ID

ROLE_ASSIGNMENT_CODE_ID

ROYALTY_CODE_ID

TEST_UTILS_CODE_ID

TRANSACTION_PROCESSOR_CODE_ID

TRANSACTION_TRACKER_CODE_ID

TRANSACTION_TRACKER_CREATE_EXPORT_NAME

TRANSACTION_TRACKER_CREATE_IDENT

VALIDATOR_PROTOCOL_VERSION_NAME_LEN

VALIDATOR_ROLE

Traits

NonFungibleResourceManagerDataContentMarker

VersionedAccessControllerStateVersion

VersionedComponentRoyaltyAccumulatorVersion

VersionedComponentRoyaltyMethodAmountVersion

VersionedConsensusManagerConfigurationVersion

VersionedConsensusManagerCurrentProposalStatisticVersion

VersionedConsensusManagerCurrentValidatorSetVersion

VersionedConsensusManagerProposerMilliTimestampVersion

VersionedConsensusManagerProposerMinuteTimestampVersion

VersionedConsensusManagerRegisteredValidatorByStakeVersion

VersionedConsensusManagerStateVersion

VersionedConsensusManagerValidatorRewardsVersion

VersionedFungibleResourceManagerDivisibilityVersion

VersionedFungibleResourceManagerTotalSupplyVersion

VersionedFungibleVaultBalanceVersion

VersionedFungibleVaultFreezeStatusVersion

VersionedFungibleVaultLockedBalanceVersion

VersionedMetadataEntryVersion

VersionedNonFungibleResourceManagerIdTypeVersion

VersionedNonFungibleResourceManagerMutableFieldsVersion

VersionedNonFungibleResourceManagerTotalSupplyVersion

VersionedNonFungibleVaultBalanceVersion

VersionedNonFungibleVaultFreezeStatusVersion

VersionedNonFungibleVaultLockedResourceVersion

VersionedNonFungibleVaultNonFungibleVersion

VersionedPackageBlueprintVersionAuthConfigVersion

VersionedPackageBlueprintVersionDefinitionVersion

VersionedPackageBlueprintVersionDependenciesVersion

VersionedPackageBlueprintVersionRoyaltyConfigVersion

VersionedPackageCodeInstrumentedCodeVersion

VersionedPackageCodeOriginalCodeVersion

VersionedPackageCodeVmTypeVersion

VersionedPackageRoyaltyAccumulatorVersion

VersionedRoleAssignmentAccessRuleVersion

VersionedRoleAssignmentOwnerVersion

VersionedValidatorProtocolUpdateReadinessSignalVersion

VersionedValidatorStateVersion

Functions

compose_proof_by_amount

Compose a proof by amount, given a list of proofs of any address.

compose_proof_by_ids

Compose a proof by ids, given a list of proofs of any address.

create_package_partition_substates

drop_fungible_bucket

drop_non_fungible_bucket

error 🔒

to_typed_object_module_substate_key

to_typed_object_substate_key_internal 🔒

to_typed_object_substate_value 🔒

to_typed_substate_key

to_typed_substate_value

to_typed_substate_value_internal 🔒

validate_metadata_value

verify_divisibility

Type Aliases

AccessControllerState

AccessControllerStateFieldSubstate

AccessControllerStateV1

ComponentClaimRoyaltiesOutput

ComponentRoyaltyAccumulator

ComponentRoyaltyAccumulatorFieldSubstate

ComponentRoyaltyAccumulatorV1

ComponentRoyaltyCreateOutput

ComponentRoyaltyLockOutput

ComponentRoyaltyMethodAmount

ComponentRoyaltyMethodAmountEntrySubstate

ComponentRoyaltyMethodAmountKeyContent

ComponentRoyaltyMethodAmountV1

ComponentRoyaltySetOutput

ConsensusManagerConfiguration

ConsensusManagerConfigurationFieldSubstate

ConsensusManagerConfigurationV1

ConsensusManagerCurrentProposalStatistic

ConsensusManagerCurrentProposalStatisticFieldSubstate

ConsensusManagerCurrentProposalStatisticV1

ConsensusManagerCurrentValidatorSet

ConsensusManagerCurrentValidatorSetFieldSubstate

ConsensusManagerCurrentValidatorSetV1

ConsensusManagerProposerMilliTimestamp

ConsensusManagerProposerMilliTimestampFieldSubstate

ConsensusManagerProposerMilliTimestampV1

ConsensusManagerProposerMinuteTimestamp

ConsensusManagerProposerMinuteTimestampFieldSubstate

ConsensusManagerProposerMinuteTimestampV1

ConsensusManagerRegisteredValidatorByStake

ConsensusManagerRegisteredValidatorByStakeEntrySubstate

ConsensusManagerRegisteredValidatorByStakeKeyContent

ConsensusManagerRegisteredValidatorByStakeV1

ConsensusManagerState

ConsensusManagerStateFieldSubstate

ConsensusManagerStateV1

ConsensusManagerValidatorRewards

ConsensusManagerValidatorRewardsFieldSubstate

ConsensusManagerValidatorRewardsV1

FungibleResourceManagerDivisibility

FungibleResourceManagerDivisibilityFieldSubstate

FungibleResourceManagerDivisibilityV1

FungibleResourceManagerTotalSupply

FungibleResourceManagerTotalSupplyFieldSubstate

FungibleResourceManagerTotalSupplyV1

FungibleVaultBalance

FungibleVaultBalanceFieldSubstate

FungibleVaultFreezeStatus

FungibleVaultFreezeStatusFieldSubstate

FungibleVaultLockedBalance

FungibleVaultLockedBalanceFieldSubstate

MetadataEntry

MetadataEntryEntrySubstate

MetadataEntryKeyContent

MetadataEntryV1

NonFungibleResourceManagerDataEntrySubstate

NonFungibleResourceManagerDataKeyContent

NonFungibleResourceManagerIdType

NonFungibleResourceManagerIdTypeFieldSubstate

NonFungibleResourceManagerIdTypeV1

NonFungibleResourceManagerMutableFields

NonFungibleResourceManagerMutableFieldsFieldSubstate

NonFungibleResourceManagerTotalSupply

NonFungibleResourceManagerTotalSupplyFieldSubstate

NonFungibleResourceManagerTotalSupplyV1

NonFungibleVaultBalance

NonFungibleVaultBalanceFieldSubstate

NonFungibleVaultFreezeStatus

NonFungibleVaultFreezeStatusFieldSubstate

NonFungibleVaultLockedResource

NonFungibleVaultLockedResourceFieldSubstate

NonFungibleVaultNonFungible

NonFungibleVaultNonFungibleEntrySubstate

NonFungibleVaultNonFungibleKeyContent

PackageBlueprintVersionAuthConfig

PackageBlueprintVersionAuthConfigEntrySubstate

PackageBlueprintVersionAuthConfigKeyContent

PackageBlueprintVersionAuthConfigV1

PackageBlueprintVersionDefinition

PackageBlueprintVersionDefinitionEntrySubstate

PackageBlueprintVersionDefinitionKeyContent

PackageBlueprintVersionDefinitionV1

PackageBlueprintVersionDependencies

PackageBlueprintVersionDependenciesEntrySubstate

PackageBlueprintVersionDependenciesKeyContent

PackageBlueprintVersionDependenciesV1

PackageBlueprintVersionRoyaltyConfig

PackageBlueprintVersionRoyaltyConfigEntrySubstate

PackageBlueprintVersionRoyaltyConfigKeyContent

PackageBlueprintVersionRoyaltyConfigV1

PackageClaimRoyaltiesOutput

PackageCodeInstrumentedCode

PackageCodeInstrumentedCodeEntrySubstate

PackageCodeInstrumentedCodeKeyContent

PackageCodeOriginalCode

PackageCodeOriginalCodeEntrySubstate

PackageCodeOriginalCodeKeyContent

PackageCodeVmType

PackageCodeVmTypeEntrySubstate

PackageCodeVmTypeKeyContent

PackagePublishNativeOutput

PackagePublishWasmAdvancedOutput

PackagePublishWasmOutput

PackageRoyaltyAccumulator

PackageRoyaltyAccumulatorFieldSubstate

PackageSchemaEntrySubstate

PackageSchemaKeyContent

RoleAssignmentAccessRule

RoleAssignmentAccessRuleEntrySubstate

RoleAssignmentAccessRuleKeyContent

RoleAssignmentAccessRuleV1

RoleAssignmentOwner

RoleAssignmentOwnerFieldSubstate

RoleAssignmentOwnerV1

TransactionStatusSubstateContents

TransactionTrackerCreateOutput

ValidatorProtocolUpdateReadinessSignal

ValidatorProtocolUpdateReadinessSignalFieldSubstate

ValidatorProtocolUpdateReadinessSignalV1

ValidatorState

ValidatorStateFieldSubstate

ValidatorStateV1